+++ lib/procon_bypass_man.rb
@@ -36,7 +36,6 @@ require_relative "procon_bypass_man/procon/button"
 require_relative "procon_bypass_man/procon/value_objects/analog_stick"
 require_relative "procon_bypass_man/procon/value_objects/procon_reader"
 require_relative "procon_bypass_man/procon/analog_stick_cap"
-require_relative "procon_bypass_man/procon/macro_builder"
 require_relative "procon_bypass_man/remote_pbm_action/value_objects/remote_pbm_action_object"
 require_relative "procon_bypass_man/scheduler"
 require_relative "procon_bypass_man/plugins"
